componentização da lista de posts e mudança na key de publication
