+++ src/pages/UserPosts.js
@@ -1,87 +1,71 @@
 import styled from "styled-components";
 import React from "react";
-import { useParams } from 'react-router-dom';
+import { useParams, useNavigate } from 'react-router-dom';
+import useInterval from "use-interval";
 
 import Header from "../components/Header";
 import SearchBar from "../components/SearchBar";
 import { getUserPostsRequest } from "../services/api";
 import ApplicationContext from "../contexts/ApplicationContext.js";
-import Publication from "../components/Publication.js";
+import PublicationList from "../components/Publication.js";
 import Trending from "../components/Trending";
 
 export default function UserPosts() {
-    const [posts, setPosts] = React.useState(null);
-    const [ notFound, setNotFound ] = React.useState(false);
+    const [posts, setPosts] = React.useState("");
+    const [numberVisiblePosts, setNumberVisiblePosts] = React.useState("");
+    const [notFound, setNotFound] = React.useState(false);
     const [username, setUsername] = React.useState("");
     const [userPic, setUserPic] = React.useState("");
     const { userToken } = React.useContext(ApplicationContext);
+    const navigate = useNavigate();
 
     const { id } = useParams()
-    
+
     const config = {
         headers: {
-          Authorization: `Bearer ${userToken}`,
+            Authorization: `Bearer ${userToken}`,
         }
     };
-    React.useEffect(() => {
-        setNotFound(false)
-        async function data(){
-            const response = getUserPostsRequest(id, config);
-            response.then( res => {
-                if (res.status === 404) {
-                    setNotFound(true)
-                } else {
-                    setUsername(res.data.userName);
-                    setUserPic(res.data.userPicUrl);
-                    setPosts([...res.data.userPosts]);
-                }
-            });
-            response.catch( err => {
-                alert("An error occured while trying to fetch the posts, please refresh the page")
-            });
-        };
-        data()
-    },[id]);
-    
-    function checkForPosts (){
-        if(posts === null){
-            return(
-                <TextContainer>
-                    <h2>Loading...</h2>
-                </TextContainer>
-            );
-        }else if(posts.length === 0){
-            return (
-                <TextContainer>
-                    <h2>There are no posts yet</h2>
-                </TextContainer>
-            );
-        }else{
-            return(
-                posts.map(item=>(
-                    <Publication  
-                        key={item.id}
-                        postId={item.id}
-                        userImage={item.pic_url}
-                        userName={item.name}
-                        authorId={item.author_id}
-                        postTitle={item.content}
-                        postLink={item.link_url}
-                        LinkName={item.link_title}
-                        LinkSummary={item.link_description}
-                        LinkImg={item.link_image}
-                        userauthorship={item.userauthorship}
-                    />))
-            );
-        };
+
+    async function data() {
+        const response = getUserPostsRequest(id, config);
+        response.then(res => {
+            if (res.status === 404) {
+                setNotFound(true)
+            } else {
+                setUsername(res.data.userName);
+                setUserPic(res.data.userPicUrl);
+                setPosts([...res.data.userPosts]);
+                setNumberVisiblePosts(res.data.userPosts.length);
+            }
+        });
+        response.catch(err => {
+            alert("An error occured while trying to fetch the posts, please refresh the page")
+        });
     };
-    const renderPosts = checkForPosts();
+
+    React.useEffect(() => {
+        if (!userToken) {
+            navigate("/", { replace: true });
+            return;
+        };
+        setNotFound(false)
+        data()
+    }, [id]);
+
+    useInterval(async () => {
+        const response = await getUserPostsRequest(id, config);
+        if (response.status !== 404) {
+            setPosts([...response.data.userPosts]);
+        }
+    }, 15000);
+
     return (
         <TimelineContainer>
             <Header />
             <SearchBar />
-            {notFound ? 
-                <h1>Usuário não encontrado</h1> 
+            {notFound ?
+                <h1>Usuário não encontrado</h1>
                 :
                 <Container>
                     <div>
@@ -89,10 +73,10 @@ export default function UserPosts() {
                             <img src={userPic} alt="user profile pic" />
                             {username}
                         </Title>
-                        {renderPosts}
+                        <PublicationList posts={posts} numberVisiblePosts={numberVisiblePosts} setNumberVisiblePosts={setNumberVisiblePosts} />
                     </div>
                     <div>
-                        <Trending /> 
+                        <Trending />
                     </div>
                 </Container>
             }
@@ -141,12 +125,3 @@ const Title = styled.h1`
         margin-left: 20px;
     }
 `;
-const TextContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    h2{
-        font: 400 20px 'Oswald', sans-serif;
-        color: #FFFFFF;
-    }
-`
